refactor(router): rename browser-navigation to browser-navigator
