+++ packages/__tests__/router/browser-navigator.spec.ts
