+++ packages/router/src/browser-navigator.ts
